requirements.txt
hidpi_tk==1.3.1
matplotlib==3.10.8
Pillow==12.1.0
weakrefset==1.0.0; python_version<"2.7"
